[Entity Store] Adopt client pattern to history_snapshot
